Extract RequiredPriceTypes to hapi-fees/ to avoid duplicating code
