+++ hedera-node/src/main/java/com/hedera/services/fees/calculation/AwareFcfsUsagePrices.java
@@ -24,6 +24,7 @@ import com.google.protobuf.InvalidProtocolBufferException;
 import com.hedera.services.config.FileNumbers;
 import com.hedera.services.context.TransactionContext;
 import com.hedera.services.files.HederaFs;
+import com.hedera.services.pricing.RequiredPriceTypes;
 import com.hederahashgraph.api.proto.java.CurrentAndNextFeeSchedule;
 import com.hederahashgraph.api.proto.java.FeeComponents;
 import com.hederahashgraph.api.proto.java.FeeData;
@@ -39,45 +40,18 @@ import org.apache.logging.log4j.Logger;
 
 import java.time.Instant;
 import java.util.EnumMap;
-import java.util.EnumSet;
-import java.util.List;
 import java.util.Map;
 import java.util.Objects;
+import java.util.Set;
 
 import static com.hedera.services.utils.EntityIdUtils.readableId;
-import static com.hederahashgraph.api.proto.java.HederaFunctionality.CryptoTransfer;
-import static com.hederahashgraph.api.proto.java.HederaFunctionality.TokenAccountWipe;
-import static com.hederahashgraph.api.proto.java.HederaFunctionality.TokenBurn;
-import static com.hederahashgraph.api.proto.java.HederaFunctionality.TokenCreate;
-import static com.hederahashgraph.api.proto.java.HederaFunctionality.TokenMint;
 import static com.hederahashgraph.api.proto.java.SubType.DEFAULT;
-import static com.hederahashgraph.api.proto.java.SubType.TOKEN_FUNGIBLE_COMMON;
-import static com.hederahashgraph.api.proto.java.SubType.TOKEN_FUNGIBLE_COMMON_WITH_CUSTOM_FEES;
-import static com.hederahashgraph.api.proto.java.SubType.TOKEN_NON_FUNGIBLE_UNIQUE;
-import static com.hederahashgraph.api.proto.java.SubType.TOKEN_NON_FUNGIBLE_UNIQUE_WITH_CUSTOM_FEES;
 
 /**
  * Implements a {@link UsagePricesProvider} by loading the required fee schedules from the Hedera "file system".
  */
 public class AwareFcfsUsagePrices implements UsagePricesProvider {
 	private static final Logger log = LogManager.getLogger(AwareFcfsUsagePrices.class);
-
-	private static final EnumSet<SubType> ONLY_DEFAULT = EnumSet.of(DEFAULT);
-	static final Map<HederaFunctionality, EnumSet<SubType>> FUNCTIONS_WITH_REQUIRED_SUBTYPES;
-	static {
-		FUNCTIONS_WITH_REQUIRED_SUBTYPES = new EnumMap<>(HederaFunctionality.class);
-		/* The functions with non-DEFAULT prices in hapi-fees/src/main/resources/canonical-prices.json */
-		List.of(TokenMint, TokenBurn, TokenAccountWipe).forEach(function ->
-				FUNCTIONS_WITH_REQUIRED_SUBTYPES.put(function, EnumSet.of(
-						TOKEN_FUNGIBLE_COMMON, TOKEN_NON_FUNGIBLE_UNIQUE)));
-		FUNCTIONS_WITH_REQUIRED_SUBTYPES.put(TokenCreate, EnumSet.of(
-				TOKEN_FUNGIBLE_COMMON, TOKEN_FUNGIBLE_COMMON_WITH_CUSTOM_FEES,
-				TOKEN_NON_FUNGIBLE_UNIQUE, TOKEN_NON_FUNGIBLE_UNIQUE_WITH_CUSTOM_FEES));
-		FUNCTIONS_WITH_REQUIRED_SUBTYPES.put(CryptoTransfer, EnumSet.of(
-				DEFAULT,
-				TOKEN_FUNGIBLE_COMMON, TOKEN_FUNGIBLE_COMMON_WITH_CUSTOM_FEES,
-				TOKEN_NON_FUNGIBLE_UNIQUE, TOKEN_NON_FUNGIBLE_UNIQUE_WITH_CUSTOM_FEES));
-	}
 
 	public static long DEFAULT_FEE = 100_000L;
 	public static final FeeComponents DEFAULT_PROVIDER_RESOURCE_PRICES = FeeComponents.newBuilder()
@@ -216,7 +190,7 @@ public class AwareFcfsUsagePrices implements UsagePricesProvider {
 			if (pricesMap == null) {
 				pricesMap = new EnumMap<>(SubType.class);
 			}
-			final EnumSet<SubType> requiredTypes = FUNCTIONS_WITH_REQUIRED_SUBTYPES.getOrDefault(function, ONLY_DEFAULT);
+			final Set<SubType> requiredTypes = RequiredPriceTypes.requiredTypesFor(function);
 			ensurePricesMapHasRequiredTypes(pricingData, pricesMap, requiredTypes);
 			allPrices.put(pricingData.getHederaFunctionality(), pricesMap);
 		}
@@ -226,7 +200,7 @@ public class AwareFcfsUsagePrices implements UsagePricesProvider {
 	void ensurePricesMapHasRequiredTypes(
 			TransactionFeeSchedule tfs,
 			Map<SubType, FeeData> pricesMap,
-			EnumSet<SubType> requiredTypes
+			Set<SubType> requiredTypes
 	) {
 		/* The deprecated prices are the final fallback; if even they are not set, the function will be free */
 		final var oldDefaultPrices = tfs.getFeeData();
