--- hedera-node/src/main/java/com/hedera/services/fees/calculation/RequiredPriceTypes.java
@@ -1,5 +1,0 @@
-package com.hedera.services.fees.calculation;
-
-public class RequiredPriceTypes {
-
-}
